refactor: extract oidc management views to module
`warehouse/manage/views/__init__.py` is already a complex module, with
over 3k LOC. Split out any OIDC code (conveniently in a single class) to
a new module.
This should allow for more focused edits when there are changes.
